+++ packages/cli/src/docker-command.ts
@@ -33,6 +33,7 @@ interface ServiceOption {
   port: number
   path: string
   p2pPort: number
+  lcdPort: number
 }
 
 export default abstract class extends Command {
@@ -150,6 +151,11 @@ export default abstract class extends Command {
           PublishMode: 'ingress',
           TargetPort: 26656,
           PublishedPort: options.p2pPort,
+        }, {
+          Protocol: 'tcp',
+          PublishMode: 'ingress',
+          TargetPort: 1317,
+          PublishedPort: options.lcdPort
         }],
       }
     })